Extract all the http and retry logic
This introduces an `HttpDriver` object that is responsible for the
lower-level interactions with the API. It handles things like URL
formats, retries, tokens, and so on. The `Client` is focused on
providing a nice interface to users and transforming the data we get
back from the API into domain objects.